# docs/swagger.yaml
basePath: /v1
definitions:
  hall.Resource:
    properties:
      ID:
        type: integer
      VIP:
        type: boolean
      seats:
        type: integer
    type: object
...
  user_privileges.Resource:
    properties:
      Email:
        type: string
      ID:
        type: integer
      Privilege:
        type: string
      Privilege_id:
        type: integer
      User_id:
        type: integer
    type: object
host: cinema-alb-dev-o81jt53c-906642332.us-east-1.elb.amazonaws.com:8085
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: This is a sample cinetickets server.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Cinetickets API
  version: "1.0"
paths:
  /halls:
    get:
      consumes:
      - application/json
      description: get halls
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              items:
                $ref: '#/definitions/hall.Resource'
              type: array
            type: array
        "400":
          description: "Bad Request"
        "401":
          description: "Unauthorized"
        "422":
          description: "Unprocessable Entity"
        "500":
          description: "Internal Server Error"
      security:
      - ApiKeyAuth: []
      summary: List halls
      tags:
      - Halls
    post:
      consumes:
      - application/json
      description: Creates hall and returns created object
      parameters:
      - description: The body to create a hall
        in: body
        name: Body
        required: true
        schema:
          $ref: '#/definitions/hall.Resource'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/hall.Resource'
        "400":
          description: "Bad Request"
        "401":
          description: "Unauthorized"
        "422":
          description: "Unprocessable Entity"
        "500":
          description: "Internal Server Error"
      security:
      - ApiKeyAuth: []
      summary: Create hall
      tags:
      - Halls
  /halls/{id}:
    delete:
      consumes:
      - application/json
      description: Deletes hall
      parameters:
      - description: Hall ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: ""
        "400":
          description: "Bad Request"
        "401":
          description: "Unauthorized"
        "422":
          description: "Unprocessable Entity"
        "500":
          description: "Internal Server Error"
      security:
      - ApiKeyAuth: []
      summary: Delete hall
      tags:
      - Halls
    get:
      consumes:
      - application/json
      description: Gets hall
      parameters:
      - description: Hall ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/hall.Resource'
        "400":
          description: "Bad Request"
        "401":
          description: "Unauthorized"
        "422":
          description: "Unprocessable Entity"
        "500":
          description: "Internal Server Error"
      security:
      - ApiKeyAuth: []
      summary: Get hall
      tags:
      - Halls
...
  /user_privileges:
    get:
      consumes:
      - application/json
      description: get User Privileges
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              items:
                $ref: '#/definitions/user_privileges.Resource'
              type: array
            type: array
        "400":
          description: "Bad Request"
        "401":
          description: "Unauthorized"
        "422":
          description: "Unprocessable Entity"
        "500":
          description: "Internal Server Error"
      security:
      - ApiKeyAuth: []
      summary: List User Privileges
      tags:
      - User Privileges
    post:
      consumes:
      - application/json
      description: Creates User Privilege and returns created object
      parameters:
      - description: The body to create a User Privilege
        in: body
        name: Body
        required: true
        schema:
          $ref: '#/definitions/user_privileges.Resource'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/user_privileges.Resource'
        "400":
          description: "Bad Request"
        "401":
          description: "Unauthorized"
        "422":
          description: "Unprocessable Entity"
        "500":
          description: "Internal Server Error"
      security:
      - ApiKeyAuth: []
      summary: Create User Privilege
      tags:
      - User Privileges
  /user_privileges/{id}:
    delete:
      consumes:
      - application/json
      description: Deletes User Privilege
      parameters:
      - description: User Privilege ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: ""
        "400":
          description: "Bad Request"
        "401":
          description: "Unauthorized"
        "422":
          description: "Unprocessable Entity"
        "500":
          description: "Internal Server Error"
      security:
      - ApiKeyAuth: []
      summary: Delete User Privilege
      tags:
      - User Privileges
    get:
      consumes:
      - application/json
      description: Gets User Privilege
      parameters:
      - description: User Privilege ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/user_privileges.Resource'
        "400":
          description: "Bad Request"
        "401":
          description: "Unauthorized"
        "422":
          description: "Unprocessable Entity"
        "500":
          description: "Internal Server Error"
      security:
      - ApiKeyAuth: []
      summary: Get User Privilege
      tags:
      - User Privileges
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"